Fixed issue in Register workflow.  Added LogIn workflow

+++ TakeHomeAssignment.Core/Controllers/Interfaces/ILogInController.cs
@@ -2,6 +2,6 @@
 {
     public interface ILogInController
     {
-        void Execute(int userId);
+        void Execute(long userId);
     }
 }
